+++ jib-gradle-plugin/src/integration-test/java/com/google/cloud/tools/jib/gradle/SingleProjectIntegrationTest.java
@@ -20,12 +20,9 @@ import com.google.cloud.tools.jib.Command;
 import com.google.cloud.tools.jib.IntegrationTestingConfiguration;
 import com.google.cloud.tools.jib.registry.LocalRegistry;
 import java.io.IOException;
-import java.nio.file.Files;
 import java.security.DigestException;
 import java.time.Instant;
 import org.gradle.testkit.runner.BuildResult;
-import org.gradle.testkit.runner.BuildTask;
-import org.gradle.testkit.runner.TaskOutcome;
 import org.gradle.testkit.runner.UnexpectedBuildFailure;
 import org.hamcrest.CoreMatchers;
 import org.junit.Assert;
@@ -178,61 +175,6 @@ public class SingleProjectIntegrationTest {
         JibRunHelper.buildToDockerDaemonAndRun(simpleTestProject, targetImage));
     assertSimpleCreationTimeIsAfter(beforeBuild, targetImage);
     assertDockerInspect(targetImage);
-  }
-
-  @Test
-  public void testDockerContext() throws IOException, InterruptedException {
-    BuildResult buildResult = simpleTestProject.build("clean", "jibExportDockerContext", "--info");
-
-    JibRunHelper.assertBuildSuccess(
-        buildResult, "jibExportDockerContext", "Created Docker context at ");
-
-    String imageName = "jib-gradle-plugin/integration-test" + System.nanoTime();
-    new Command(
-            "docker",
-            "build",
-            "-t",
-            imageName,
-            simpleTestProject
-                .getProjectRoot()
-                .resolve("build")
-                .resolve("jib-docker-context")
-                .toString())
-        .run();
-
-    assertDockerInspect(imageName);
-    String output = new Command("docker", "run", "--rm", imageName).run();
-    Assert.assertThat(output, CoreMatchers.startsWith("Hello, world. An argument.\n"));
-    Assert.assertThat(output, CoreMatchers.endsWith("foo\ncat\n"));
-
-    // Checks that generating the Docker context again is skipped.
-    BuildTask upToDateJibDockerContextTask =
-        simpleTestProject.build("jibExportDockerContext").task(":jibExportDockerContext");
-    Assert.assertNotNull(upToDateJibDockerContextTask);
-    Assert.assertEquals(TaskOutcome.UP_TO_DATE, upToDateJibDockerContextTask.getOutcome());
-
-    // Checks that adding a new file generates the Docker context again.
-    Files.createFile(
-        simpleTestProject
-            .getProjectRoot()
-            .resolve("src")
-            .resolve("main")
-            .resolve("resources")
-            .resolve("newfile"));
-    try {
-      BuildTask reexecutedJibDockerContextTask =
-          simpleTestProject.build("jibExportDockerContext").task(":jibExportDockerContext");
-      Assert.assertNotNull(reexecutedJibDockerContextTask);
-      Assert.assertEquals(TaskOutcome.SUCCESS, reexecutedJibDockerContextTask.getOutcome());
-
-    } catch (UnexpectedBuildFailure ex) {
-      // This might happen on systems without SecureDirectoryStream, so we just ignore it.
-      // See com.google.common.io.MoreFiles#deleteDirectoryContents.
-      Assert.assertThat(
-          ex.getMessage(),
-          CoreMatchers.containsString(
-              "Export Docker context failed because cannot clear directory"));
-    }
   }
 
   @Test