+++ jib-gradle-plugin/src/main/java/com/google/cloud/tools/jib/gradle/ContainerParameters.java
@@ -74,6 +74,10 @@ public class ContainerParameters {
     this.entrypoint = entrypoint;
   }
 
+  public void setEntrypoint(String entrypoint) {
+    this.entrypoint = Collections.singletonList(entrypoint);
+  }
+
   @Input
   @Optional
   public List<String> getJvmFlags() {